requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0
